Don't set an end date for the current term
The way we generate term dates is to look for the earliest and latest
dates of memberships in that term. But that means that we set the end
date of the current term to the latest end_date of any Membership in it.
Instead we should explicitly delete that date.